+++ tests/conftest.py
@@ -9,6 +9,7 @@ except ImportError:
 
 import pytest
 
+from configurence import BaseConfig
 from configurence import config as config_
 from configurence import field
 
@@ -26,9 +27,9 @@ def convert_other(value: str) -> Other:
 
 
 @pytest.fixture
-def config_cls():
-    @config_
-    class Config:
+def config_cls(app_name):
+    @config_(app_name)
+    class Config(BaseConfig):
         some_str: str
         opt_str: Optional[str]
         some_bool: bool
@@ -47,19 +48,18 @@ def config_cls():
 
 @pytest.fixture
 def app_name() -> str:
-    return "my-app"
+    return "test-app"
 
 
 @pytest.fixture
-def local_filename() -> str:
-    return "/Users/josh/.config/my-app.yaml"
+def local_filename(app_name) -> str:
+    return f"/Users/josh/.config/{app_name}.yaml"
 
 
 @pytest.fixture
-def config(config_cls, app_name, local_filename):
+def config(config_cls, local_filename):
     return config_cls(
-        name=app_name,
-        _file=local_filename,
+        file=local_filename,
         some_str="some_str",
         opt_str=None,
         some_bool=True,