+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-yyyy/mm/dd Version x.y.z
+2025/01/23 Version 2.0.0
 ------------------------
-- TK
+- Breaking API changes:
+  - Must inherit from `BaseConfig`
+  - App name set by `@config` decorator
+
+2025/01/23 Version 1.0.0
+------------------------
+- First draft